Move file events to their own package
