+++ src/main/java/net/rubygrapefruit/platform/internal/Platform.java
@@ -27,13 +27,13 @@ import net.rubygrapefruit.platform.file.FileSystems;
 import net.rubygrapefruit.platform.file.Files;
 import net.rubygrapefruit.platform.file.PosixFiles;
 import net.rubygrapefruit.platform.file.WindowsFiles;
-import net.rubygrapefruit.platform.internal.jni.AbstractFileEventFunctions;
-import net.rubygrapefruit.platform.internal.jni.LinuxFileEventFunctions;
 import net.rubygrapefruit.platform.internal.jni.NativeVersion;
-import net.rubygrapefruit.platform.internal.jni.OsxFileEventFunctions;
 import net.rubygrapefruit.platform.internal.jni.PosixTypeFunctions;
 import net.rubygrapefruit.platform.internal.jni.TerminfoFunctions;
-import net.rubygrapefruit.platform.internal.jni.WindowsFileEventFunctions;
+import net.rubygrapefruit.platform.internal.jni.fileevents.AbstractFileEventFunctions;
+import net.rubygrapefruit.platform.internal.jni.fileevents.LinuxFileEventFunctions;
+import net.rubygrapefruit.platform.internal.jni.fileevents.OsxFileEventFunctions;
+import net.rubygrapefruit.platform.internal.jni.fileevents.WindowsFileEventFunctions;
 import net.rubygrapefruit.platform.memory.Memory;
 import net.rubygrapefruit.platform.memory.OsxMemory;
 import net.rubygrapefruit.platform.terminal.Terminals;
