+++ src/main/java/net/rubygrapefruit/platform/internal/jni/fileevents/LinuxFileEventFunctions.java
@@ -1,20 +1,20 @@
 /*
- * Copyright 2012 Adam Murdoch
+ * Copyright 2020 the original author or authors.
  *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *        http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
-package net.rubygrapefruit.platform.internal.jni;
+package net.rubygrapefruit.platform.internal.jni.fileevents;
 
 import net.rubygrapefruit.platform.file.FileWatcher;
 import net.rubygrapefruit.platform.file.FileWatcherCallback;
@@ -28,7 +28,7 @@ public class LinuxFileEventFunctions extends AbstractFileEventFunctions {
      *
      * <ul>
      *     <li>Changes to descendants to the given paths are not reported.</li>
-     * 
+     *
      *     <li>TBD</li>
      * </ul>
      */
