+++ src/main/java/net/rubygrapefruit/platform/internal/jni/fileevents/NativeLogger.java
@@ -1,4 +1,20 @@
-package net.rubygrapefruit.platform.internal.jni;
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.rubygrapefruit.platform.internal.jni.fileevents;
 
 import java.util.logging.Level;
 import java.util.logging.Logger;
