+++ src/test/groovy/net/rubygrapefruit/platform/internal/jni/fileevents/BasicFileEventFunctionsTest.groovy
@@ -1,24 +1,23 @@
 /*
- * Copyright 2012 Adam Murdoch
+ * Copyright 2020 the original author or authors.
  *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *        http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-package net.rubygrapefruit.platform.file
+package net.rubygrapefruit.platform.internal.jni.fileevents
 
 import net.rubygrapefruit.platform.NativeException
+import net.rubygrapefruit.platform.file.FileWatcherCallback
 import net.rubygrapefruit.platform.internal.Platform
-import net.rubygrapefruit.platform.internal.jni.AbstractFileEventFunctions
-import net.rubygrapefruit.platform.internal.jni.NativeLogger
 import org.junit.Assume
 import spock.lang.Ignore
 import spock.lang.IgnoreIf
