+++ test-app/src/main/java/net/rubygrapefruit/platform/test/Main.java
@@ -33,9 +33,9 @@ import net.rubygrapefruit.platform.file.Files;
 import net.rubygrapefruit.platform.file.PosixFileInfo;
 import net.rubygrapefruit.platform.file.PosixFiles;
 import net.rubygrapefruit.platform.internal.Platform;
-import net.rubygrapefruit.platform.internal.jni.LinuxFileEventFunctions;
-import net.rubygrapefruit.platform.internal.jni.OsxFileEventFunctions;
-import net.rubygrapefruit.platform.internal.jni.WindowsFileEventFunctions;
+import net.rubygrapefruit.platform.internal.jni.fileevents.LinuxFileEventFunctions;
+import net.rubygrapefruit.platform.internal.jni.fileevents.OsxFileEventFunctions;
+import net.rubygrapefruit.platform.internal.jni.fileevents.WindowsFileEventFunctions;
 import net.rubygrapefruit.platform.memory.Memory;
 import net.rubygrapefruit.platform.memory.MemoryInfo;
 import net.rubygrapefruit.platform.prompts.Prompter;
